refactor: expose curated feature entrypoints
