+++ app/frontend/src/composables/compose/useAdapterCatalog.ts
@@ -1,7 +1,7 @@
 import { computed, ref, type ComputedRef, type Ref } from 'vue';
 import { storeToRefs } from 'pinia';
 
-import { useAdapterCatalogStore } from '@/features/lora';
+import { useAdapterCatalogStore } from '@/features/lora/public';
 
 import type { AdapterSummary } from '@/types';
