+++ app/frontend/src/features/analytics/public.ts
@@ -1,0 +1,9 @@
+export { default as PerformanceAnalyticsChartGrid } from './components/PerformanceAnalyticsChartGrid.vue';
+export { default as PerformanceAnalyticsExportToolbar } from './components/PerformanceAnalyticsExportToolbar.vue';
+export { default as PerformanceAnalyticsInsights } from './components/PerformanceAnalyticsInsights.vue';
+export { default as PerformanceAnalyticsKpiGrid } from './components/PerformanceAnalyticsKpiGrid.vue';
+
+export { usePerformanceAnalytics } from './composables/usePerformanceAnalytics';
+
+export { usePerformanceAnalyticsStore } from './stores/performanceAnalytics';
+export type { PerformanceAnalyticsStore } from './stores/performanceAnalytics';
