+++ app/frontend/src/features/generation/components/GenerationActiveJobsList.vue
@@ -69,7 +69,7 @@
 <script setup lang="ts">
 import { toRefs } from 'vue'
 
-import type { UseGenerationStudioReturn } from '@/features/generation'
+import type { UseGenerationStudioReturn } from '../composables/useGenerationStudio'
 import type { GenerationJob } from '@/types'
 
 const props = defineProps<{
