+++ app/frontend/src/features/generation/index.ts
@@ -1,4 +1,9 @@
-export * from './components';
-export * from './composables';
-export * from './services';
-export * from './stores';
+export {
+  GenerationStudio,
+  JobQueue,
+  SystemAdminStatusCard,
+  SystemStatusCard,
+  SystemStatusPanel,
+  useGenerationResultsStore,
+} from './public';
+export type { GenerationResultsStore } from './public';
