+++ app/frontend/src/features/lora/components/lora-gallery/LoraGallery.vue
@@ -69,9 +69,9 @@ import LoraGalleryFilters from './LoraGalleryFilters.vue';
 import LoraGalleryGrid from './LoraGalleryGrid.vue';
 import LoraGalleryHeader from './LoraGalleryHeader.vue';
 import LoraGalleryTagModal from './LoraGalleryTagModal.vue';
-import { useLoraGalleryData } from '@/features/lora/composables';
-import { useLoraGalleryFilters } from '@/features/lora/composables';
-import { useLoraGallerySelection } from '@/features/lora/composables';
+import { useLoraGalleryData } from '../../composables/lora-gallery/useLoraGalleryData';
+import { useLoraGalleryFilters } from '../../composables/lora-gallery/useLoraGalleryFilters';
+import { useLoraGallerySelection } from '../../composables/lora-gallery/useLoraGallerySelection';
 import { useDialogService, useNotifications, useSyncedQueryParam } from '@/composables/shared';
 import type {
   LoraBulkAction,
