+++ app/frontend/src/features/lora/index.ts
@@ -1,4 +1,2 @@
-export * from './components';
-export * from './composables';
-export * from './services';
-export * from './stores';
+export { LoraGallery, useAdapterCatalogStore, fetchTopAdapters } from './public';
+export type { AdapterCatalogStore } from './public';
