+++ app/frontend/src/features/lora/public.ts
@@ -1,0 +1,6 @@
+export { default as LoraGallery } from './components/lora-gallery/LoraGallery.vue';
+
+export { useAdapterCatalogStore } from './stores/adapterCatalog';
+export type { AdapterCatalogStore } from './stores/adapterCatalog';
+
+export { fetchTopAdapters } from './services/lora/loraService';
