+++ app/frontend/src/views/AdminView.vue
@@ -31,7 +31,7 @@ import { RouterLink } from 'vue-router';
 import { defineAsyncComponent } from 'vue';
 
 import ImportExportContainer from '@/components/import-export/ImportExportContainer.vue';
-import { JobQueue, SystemAdminStatusCard, SystemStatusPanel } from '@/features/generation';
+import { JobQueue, SystemAdminStatusCard, SystemStatusPanel } from '@/features/generation/public';
 import PageHeader from '@/components/layout/PageHeader.vue';
 import { RecommendationsPanel } from '@/features/recommendations';
