+++ app/frontend/src/views/DashboardView.vue
@@ -74,10 +74,15 @@ import { RouterLink } from 'vue-router';
 import DashboardGenerationSummary from '@/components/dashboard/DashboardGenerationSummary.vue';
 import DashboardLazyModuleCard from '@/components/dashboard/DashboardLazyModuleCard.vue';
 import DashboardLoraSummary from '@/components/dashboard/DashboardLoraSummary.vue';
-import { JobQueue, SystemAdminStatusCard, SystemStatusCard, SystemStatusPanel } from '@/features/generation';
+import {
+  JobQueue,
+  SystemAdminStatusCard,
+  SystemStatusCard,
+  SystemStatusPanel,
+} from '@/features/generation/public';
 import PageHeader from '@/components/layout/PageHeader.vue';
 import { RecommendationsPanel } from '@/features/recommendations';
-import { usePerformanceAnalyticsStore } from '@/features/analytics';
+import { usePerformanceAnalyticsStore } from '@/features/analytics/public';
 
 type PanelKey = 'analytics' | 'composer' | 'studio' | 'gallery' | 'history' | 'importExport';
