+++ app/frontend/src/views/HistoryView.vue
@@ -24,6 +24,6 @@
 import { RouterLink } from 'vue-router';
 
 import { GenerationHistory } from '@/features/history';
-import { JobQueue, SystemStatusCard } from '@/features/generation';
+import { JobQueue, SystemStatusCard } from '@/features/generation/public';
 import PageHeader from '@/components/layout/PageHeader.vue';
 </script>
